+++ CHANGELOG.md
@@ -10,3 +10,5 @@ The Job Portal App Changes
   [FEATURE] App: Build the UI with Mock Data
 - [JBM-7](https://shahidpj06.atlassian.net/browse/JBM-7)
   [FEATURE] Server: Built the Database Foundation: PostgreSQL, Prisma Schema, and Seed Data
+- [JBM-8](https://shahidpj06.atlassian.net/browse/JBM-8)
+  [FEATURE] Server: Authentication API: JWT, Refresh Tokens, and Authorization